Fix a typo
The handout asks to use the `map` or `switchMap` transformations methods of the `LiveData` class. This is a mistake since there is no `map`/`switchMap` methods on the `LiveData` class.

Note as well that those methods are static.

See [Android's LiveData documentation](https://developer.android.com/topic/libraries/architecture/livedata#livedata-in-architecture).